Features/fix race condition 2 (#80)
Took a while to debug this bugger.

The mix of async, Q and standard promise was causing all kinds of funny. It turned out to be easier to rewrite the code and make it a lot neater anyway. All tests are looking good, but this is a quite impactful change.

Fixes #78