Update UnpublishExtension.ts
Hey,
When testing the new 'unpublish' feature, we got a 401 error. When double-checking the pipeline, I noticed that the 'service-url' parameter was missing.
I also checked the code and I notice that the function common.setTfxMarketplaceArguments(tfx, false); is been called with the false parameter.
I am wondering if this is related to the issue.
Thank you!