Always set both local and output variables for `Extension.Version`
Found the culprit of this bug.

When the task.json explicitly lists the output variable, the variable is *always* promoted to an output variable.
When the task.json does not explicitly list the output variable, one can set *both* the local and the output variable.

This seems like a useful compromise, since the output variable isn't delcaration isn't shown by the YAML step editor UI anyway.